충남대 BE_신성희 6주차 과제 (3단계) (#376)
* docs : README 작성

* feat : 사용자가 포인트를 보유하도록 member entity 수정

* feat : 포인트 충전, 조회 기능 구현 - token으로 어떤 멤버의 포인트를 충전하고 조회할지 확인 하고 반영후 , 멤버가 가진 최종 포인트를 반환

* feat : 이전 커밋에 Service 부분 안올려서 다시올림
- 포인트 충전, 조회 기능 구현

* fix : 멤버 포인트 충전 HTTP 메서드 수정 (팀활동에서 정한걸로 수정)

* feat : 포인트 차감 기능 구현

* feat : gradle 버전 변경 17 -> 21
- Swagger 등 작성한 코드들 때문에 21로 변경하였습니다.

* fix : 회원 가입 ( 회원등록) 기능 안되던 문제 해결

+++ src/main/java/gift/entity/Product.java
@@ -29,7 +29,7 @@ public class Product {
     private String name;
     @Column(nullable = false)
     @Schema(description = "상품 가격")
-    private Integer price;
+    private Long price;
     @Column(nullable = false)
     @Schema(description = "상품 url")
     private String url;
@@ -44,7 +44,7 @@ public class Product {
     public Product() {
     }
 
-    public Product(String name, Integer price, String url, Category category, List<Option> options) {
+    public Product(String name, Long price, String url, Category category, List<Option> options) {
         this.name = name;
         this.price = price;
         this.url = url;
@@ -52,7 +52,7 @@ public class Product {
         this.options = options;
     }
 
-    public Product(Long id, String name, Integer price, String url, Category category, List<Option> options) {
+    public Product(Long id, String name, Long price, String url, Category category, List<Option> options) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -61,7 +61,7 @@ public class Product {
         this.options = options;
     }
 
-    public Product(String name, Integer price, String url) {
+    public Product(String name, Long price, String url) {
         this.name = name;
         this.price = price;
         this.url = url;
@@ -75,7 +75,7 @@ public class Product {
         return name;
     }
 
-    public Integer getPrice() {
+    public Long getPrice() {
         return price;
     }
 
@@ -91,7 +91,7 @@ public class Product {
         return options;
     }
 
-    public void update(String name, Integer price, String url, Category category, List<Option> options) {
+    public void update(String name, Long price, String url, Category category, List<Option> options) {
         this.name = name;
         this.price = price;
         this.url = url;
